+++ packages/canary-bot/src/canary-bot.ts
@@ -13,9 +13,11 @@
 // limitations under the License.
 
 /* eslint-disable-next-line node/no-extraneous-import */
-import {Probot} from 'probot';
+import {Probot, Logger} from 'probot';
 /* eslint-disable-next-line node/no-extraneous-import */
 import {components} from '@octokit/openapi-types';
+/* eslint-disable-next-line node/no-extraneous-import */
+import {Octokit} from '@octokit/rest';
 import * as fs from 'fs';
 import {resolve} from 'path';
 import {logger, addOrUpdateIssueComment} from 'gcf-utils';
@@ -40,6 +42,14 @@ const cronIssueTitle = 'A canary is chirping';
 const myRepositoryName = 'repo-automation-bots';
 const myOrganizationName = 'googleapis';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+interface PubSubContext<T = any> {
+  github: Octokit;
+  readonly event: string;
+  log: Logger;
+  payload: T;
+}
+
 function getIssueBody(): string {
   const date = dayjs
     .tz(new Date(), 'America/Los_Angeles')
@@ -51,7 +61,8 @@ function getIssueBody(): string {
 }
 
 export = (app: Probot) => {
-  app.on('schedule.repository' as '*', async context => {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  app.on('schedule.repository' as any, async context => {
     const owner = context.payload.organization.login;
     const repo = context.payload.repository.name;
     if (repo !== myRepositoryName || owner !== myOrganizationName) {
@@ -90,13 +101,15 @@ export = (app: Probot) => {
     }
   });
 
-  app.on('schedule.installation' as '*', async context => {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  app.on('schedule.installation' as any, async context => {
     logger.info(
       `executed scheduled task for installation: ${context.payload.installation.id}`
     );
   });
 
-  app.on('schedule.global' as '*', async () => {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  app.on('schedule.global' as any, async () => {
     logger.info('executed global scheduled task');
   });
 
@@ -116,5 +129,14 @@ export = (app: Probot) => {
         'The bot is skipping this issue because the title does not include canary-bot test'
       );
     }
+  });
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  app.on('pubsub.message' as any, async context => {
+    const pubsubContext = context as unknown as PubSubContext;
+    logger.info(
+      'executed pubsub handler with the payload: ' +
+        `${JSON.stringify(pubsubContext.payload)}`
+    );
   });
 };