+++ packages/release-please/src/config-constants.ts
@@ -35,6 +35,7 @@ export interface BranchConfiguration extends BranchOptions {
 export interface ConfigurationOptions extends BranchOptions {
   primaryBranch: string;
   branches?: BranchConfiguration[];
+  disableFailureChecker?: boolean;
 }
 
 export const WELL_KNOWN_CONFIGURATION_FILE = 'release-please.yml';